___JmolVersion="13.1.15_dev_2013.04.24b"

+++ trunk/src/org/jmol/render/SticksRenderer.java
@@ -1,7 +1,7 @@
 /* $RCSfile$
  * $Author: hansonr $
- * $Date: 2013-04-14 18:18:39 -0500 (Sun, 14 Apr 2013) $
- * $Revision: 18110 $
+ * $Date: 2013-04-21 11:52:35 -0500 (Sun, 21 Apr 2013) $
+ * $Revision: 18136 $
 
  *
  * Copyright (C) 2002-2005  The Jmol Development Team
@@ -251,7 +251,7 @@ public class SticksRenderer extends ShapeRenderer {
       mad *= multipleBondRadiusFactor;
     dx = xB - xA;
     dy = yB - yA;
-    width = viewer.scaleToScreen((zA + zB) / 2, mad);
+    width = (int) viewer.scaleToScreen((zA + zB) / 2, mad);
     if (renderWireframe && width > 0)
       width = 1;
     if (!isCartesianExport) {
